making quatinty for umutiSold to be float instead of int.

+++ pharma/models.py
@@ -160,7 +160,7 @@ class UmutiSold(models.Model):
     """This one will record all the sale and benefit as well"""
     code_med = models.CharField(max_length=8, default='null')
     nom_med = models.CharField(max_length=30, default='null')
-    quantity = models.IntegerField(default=1) #quantity sold
+    quantity = models.FloatField(default=1) #quantity sold
     prix_vente = models.IntegerField(default=0) #unit price
     price_total = models.IntegerField(default=1) # q * p
     prix_achat = models.IntegerField(default=1)
